Fix `non-serializable value in the state` error
There was a warning in the console - `non-serializable value was detected in the
state`. This happened because we passed two functions (`handleJsonDownload` and
`handleDoubleReject`) through the `openModal` action and function are
non-seriazible.
